Created PortableUtil shared DLL and moved ConsoleTable into it
* Added pwiz_tools/Shared/PortableUtil (SDK, net472;net8.0, pure-BCL leaf)
* Moved ConsoleTable to pwiz.Common.CommandLine namespace, made it public
* Replaced ConsoleTable's Assume.IsTrue with an InvalidOperationException throw
* Wired Skyline ProjectReference and Skyline.sln config rows (x86->AnyCPU)

See ai/todos/active/TODO-20260622_portableutil_cli_args_framework.md
