+++ pwiz_tools/Skyline/CommandArgs.cs
@@ -29,6 +29,7 @@ using System.Web;
 using System.Xml.Serialization;
 using pwiz.Common.Chemistry;
 using pwiz.Common.Collections;
+using pwiz.Common.CommandLine;
 using pwiz.Common.DataBinding.Documentation;
 using pwiz.Common.SystemUtil;
 using pwiz.CommonMsData;
