+++ pwiz_tools/Skyline/Skyline.sln
@@ -35,6 +35,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "..\Shared\Common\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonUtil", "..\Shared\CommonUtil\CommonUtil.csproj", "{13BF2FFB-50A1-4AB1-83A4-5733E36905CE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PortableUtil", "..\Shared\PortableUtil\PortableUtil.csproj", "{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BiblioSpec", "..\Shared\BiblioSpec\BiblioSpec.csproj", "{32CC9B1A-9442-4B56-AF34-FB47595278A1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestRunner", "TestRunner\TestRunner.csproj", "{B92CA24E-0867-4B99-88F8-EF542A4AFE87}"
@@ -177,6 +179,14 @@ Global
 		{13BF2FFB-50A1-4AB1-83A4-5733E36905CE}.Release|x64.Build.0 = Release|x64
 		{13BF2FFB-50A1-4AB1-83A4-5733E36905CE}.Release|x86.ActiveCfg = Release|x86
 		{13BF2FFB-50A1-4AB1-83A4-5733E36905CE}.Release|x86.Build.0 = Release|x86
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Debug|x64.ActiveCfg = Debug|x64
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Debug|x64.Build.0 = Debug|x64
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Debug|x86.Build.0 = Debug|Any CPU
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Release|x64.ActiveCfg = Release|x64
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Release|x64.Build.0 = Release|x64
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Release|x86.ActiveCfg = Release|Any CPU
+		{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}.Release|x86.Build.0 = Release|Any CPU
 		{32CC9B1A-9442-4B56-AF34-FB47595278A1}.Debug|x64.ActiveCfg = Debug|x64
 		{32CC9B1A-9442-4B56-AF34-FB47595278A1}.Debug|x64.Build.0 = Debug|x64
 		{32CC9B1A-9442-4B56-AF34-FB47595278A1}.Debug|x86.ActiveCfg = Debug|x86
